pyproject.toml
[project]
name = "jobde-ai-job-search-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncio>=4.0.0",
    "groq>=0.36.0",
    "instructor>=1.13.0",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langgraph>=1.0.3",
    "langgraph-checkpoint>=3.0.1",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
